// Frontend/quran-dashboard-ui/src/app/core/api/generated/models/access-user-summary.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface AccessUserSummary {
  createdAtUtc: string;
  displayName: string | null;
  email: string;
  id: number;
  isOwner: boolean;
  permissionCount: number;
  status: string;
  updatedAtUtc: string;
  version: number;
}